modified:   a.out 	new file:   background-stuff/fixing-prevent.cpp 	renamed:    fixing_diag.cpp -> background-stuff/fixing_diag.cpp 	new file:   diag.h 	modified:   diagnostic.cpp 	new file:   headers.h 	modified:   main.cpp 	new file:   prevent.h 	modified:   preventative.cpp
