+++ background-stuff/fixing_diag.cpp
