+++ headers.h
@@ -1,0 +1,8 @@
+#include <cctype>
+#include <fstream>
+#include <iostream>
+#include <string>
+//#include "diagnostic.cpp"
+//#include "preventative.cpp"
+#include "diag.h"
+#include "prevent.h"
