+++ vehicle-directory/4231.txt4213.txt
@@ -1,0 +1,4 @@
+Be aware: the transmission is slipping.
+Issue: dirty or burnt transmission fluid. Order: transmission fluid flush.
+Diagnostic report for 4213 vehicle:
+Issue reported: Transmission
